requirements.txt
requests
pandas
flask
sqlalchemy
psycopg2-binary